requirements.txt
Pillow==11.0.0
torchvision==0.16.2
transformers==4.44.0
